# product-service/src/functions/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
      "description": "This is a product service documentation",
      "version": "1.0.0",
      "title": "Product service",
      "license": {
        "name": "Apache 2.0",
        "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
      }
    },
    "host": "jfrp9feove.execute-api.us-east-1.amazonaws.com",
    "basePath": "/dev",
    "tags": [
      {
        "name": "product",
        "description": "Everything about your Pets"
      }
    ],
    "schemes": [
      "https"
    ],
    "paths": {
      "/products": {
        "get": {
          "tags": [
            "product"
          ],
          "summary": "Get all products",
          "description": "Get all products",
          "operationId": "getProductsList",
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": "successful operation",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/Product"
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "product"
          ],
          "summary": "Create product",
          "description": "Create product",
          "operationId": "createProduct",
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "body",
              "in": "body",
              "description": "Product to create",
              "required": true,
              "schema": {
                "$ref": "#/definitions/Product"
              }
            }
          ],
          "responses": {
            "201": {
              "description": "successful operation",
              "schema": {
                "$ref": "#/definitions/Product"
              }
            },
            "400": {
              "description": "invalid request body"
            }
          }
        }
      },
      "/products/{id}": {
        "get": {
          "tags": [
            "product"
          ],
          "summary": "Find product by ID",
          "description": "Returns a single product",
          "operationId": "getProductById",
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID of product to return",
              "required": true,
              "type": "string",
              "format": "uuid"
            }
          ],
          "responses": {
            "200": {
              "description": "successful operation",
              "schema": {
                "$ref": "#/definitions/Product"
              }
            },
            "404": {
              "description": "Product not found"
            }
          }
        }
      }
    },
    "definitions": {
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "price": {
            "type": "integer",
            "format": "int64"
          },
          "count": {
            "type": "integer",
            "format": "int64"
          }
        },
        "xml": {
          "name": "Product"
        }
      }
    },
    "externalDocs": {
      "description": "Find out more about Swagger",
      "url": "http://swagger.io"
    }
  }